feat(sys): remove init.tar support
The switch I want to make is to create an init program in C which
gets embedded without any file system abstraction like tar.

It is a simpler and more efficient way to load it
to RAM than embedding a tar archive. I would have not used the tar
file system anyways, I think.